+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog of lizard-nxt client
 unreleased
 -----------
 
+- Events with start and end as line in timeline
+
+- Cumulative rain for spatial extent in timeline
+
 - Event aggregate table in object cards (Performance Indicator)
 
 - Event aggregate table for eventseries (Performance Indicator)